requirements.txt
numpy==2.1.3
PyYAML==6.0.2
PyQt6==6.8.1
PyOpenGL==3.1.7
pytest==8.3.5
